+++ client/CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Abort the sync nodes task through its handle instead of a oneshot channel;
+- Update network info also from unhealty nodes if ignore_node_health is set to true;
 
 ### Removed